refactor(files): move alt_bn128 to altbn_128
