+++ src/lib.rs
@@ -1,6 +1,6 @@
 extern crate alloc;
 
-pub mod alt_bn128;
+pub mod altbn_128;
 pub mod bls12_381;
 pub mod scratch;
