+++ src/scratch.rs
@@ -89,49 +89,3 @@ pub trait Curve: Engine + ScalarEngine {
 }
 
 impl<T> Curve for T where T: Engine + ScalarEngine {}
-
-/// Pairing-Friendly Curve ALTBN128
-pub mod altbn_128 {
-    use super::Curve;
-    use pairing_ce::{bn256::Bn256, GroupDecodingError};
-
-    const G1_LENGTH: usize = 96;
-
-    /// altbn_128 add
-    pub fn add(input: &[u8], output: &mut [u8]) -> Result<(), GroupDecodingError> {
-        <Bn256 as Curve>::add(input, output)
-    }
-
-    /// altbn_128 mul
-    pub fn mul(input: &[u8], output: &mut [u8]) -> Result<(), GroupDecodingError> {
-        <Bn256 as Curve>::mul(input, output)
-    }
-
-    /// altbn_128 pairing
-    pub fn pairing(input: &[u8]) -> Result<bool, GroupDecodingError> {
-        <Bn256 as Curve>::pairing(input, G1_LENGTH)
-    }
-}
-
-/// Pairing-Friendly Curve BLS12_381
-pub mod bls12_381 {
-    use super::Curve;
-    use pairing_ce::{bls12_381::Bls12, GroupDecodingError};
-
-    const G1_LENGTH: usize = 64;
-
-    /// altbn_128 add
-    pub fn add(input: &[u8], output: &mut [u8]) -> Result<(), GroupDecodingError> {
-        <Bls12 as Curve>::add(input, output)
-    }
-
-    /// altbn_128 mul
-    pub fn mul(input: &[u8], output: &mut [u8]) -> Result<(), GroupDecodingError> {
-        <Bls12 as Curve>::mul(input, output)
-    }
-
-    /// altbn_128 pairing
-    pub fn pairing(input: &[u8]) -> Result<bool, GroupDecodingError> {
-        <Bls12 as Curve>::pairing(input, G1_LENGTH)
-    }
-}
